add PTI original (github.com/danielroich/PTI)

+++ utils/ImagesDataset.py
@@ -25,7 +25,7 @@ class ImagesDataset(Dataset):
         from_im = from_im.crop(((w - s) // 2, (h - s) // 2, (w + s) // 2, (h + s) // 2))
         if self.dezired_size<=0: self.dezired_size=s
         from_im = from_im.resize((self.dezired_size, self.dezired_size), Image.LANCZOS)
-        from_im = np.array(from_im, dtype=np.uint8)
+        from_im = np.array(from_im, dtype=np.float32)
         
         if self.source_transform:
             from_im = self.source_transform(from_im)
